Update health endpoint to status
