+++ packages/common/src/service/Health/Health.ts
@@ -1,3 +1,4 @@
-export type Version = {
+export type APIStatus = {
+	status: string;
 	version: string;
 }
